requirements.txt
attrs==24.2.0
cattrs==24.1.2
certifi==2024.8.30
charset-normalizer==3.4.0
cssselect==1.2.0
idna==3.10
jmespath==1.0.1
lxml==5.3.0
packaging==24.2
parsel==1.9.1
platformdirs==4.3.6
PoroPilot==0.2.10
python-dotenv==1.0.1
requests==2.32.3
requests-cache==1.2.1
six==1.16.0
url-normalize==1.4.3
urllib3==2.2.3
w3lib==2.2.1
